moved notes into user-page auto fetch applied
