+++ code/client/src/App.tsx
@@ -1,4 +1,4 @@
-import { UserPage } from "./userpage"
+import { UserPage } from "./user-page"
 import { Users } from "./users";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
